pyproject.toml
[project]
name = "email-sender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtpd>=1.4.6",
    "jinja2>=3.1.6",
    "mjml-python>=1.3.7",
    "openai-agents[litellm]>=0.6.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
